==> notebooks/03_featuring.ipynb <==
## 00 - importation librairies et dataframe

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

from utils import dataframe_info, racine_projet

racine_projet()

df = pd.read_csv(racine_projet()+'/data/raw/merged_data.csv', low_memory = False)

## 01 - data cleaning pre-viz

#### Création de la Target : ResponseDuration

df['DateAndTimeArrived'] = pd.to_datetime(df['DateAndTimeArrived'])
df['DateAndTimeMobilised'] = pd.to_datetime(df['DateAndTimeMobilised'])

df['ResponseDuration'] = df['DateAndTimeArrived'] - df['DateAndTimeMobilised']
df['ResponseDuration'] = df['ResponseDuration'].abs().dt.total_seconds().astype(int)

#### Création de la colonne DateAndTimeCalled pour supprimer les colonnes :
DateOfCall / CalYear / TimeOfCall / HourOfCall 

df['DateAndTimeCalled']= df['DateOfCall']+ [' '] + df['TimeOfCall']

#### Création de la colonne IncidentType


# Créer la nouvelle variable 'IncidentType'
df['IncidentType'] = df['StopCodeDescription']

# Remplacer les valeurs 'special Service' par le contenu de 'SpecialCodeDescription'
df.loc[df['StopCodeDescription'] == 'Special Service', 'IncidentType'] = df['SpecialServiceType']

#### Suppression de variables

#Suppresion pour cause d'inutilité / variables administratives
variables_drop0 =['IncidentNumber',
                  'ResourceMobilisationId',
                  'Resource_Code',
                  'FRS']
df_dropped = df.drop(labels = variables_drop0, axis = 1)

#Suppresion pour cause de doublon/ redondance par codification
variables_drop1 =['ProperCase',
                  'DeployedFromStation_Code',
                  'PlusCode_Code',
                  'DelayCodeId',
                  'ProperCase',
                  'PerformanceReporting',
                  'IncGeo_WardCode',
                  'IncGeo_BoroughCode']
df_dropped = df_dropped.drop(labels = variables_drop1, axis = 1)

# Suppression pour cause de %NAN trop élevé
variables_drop2 =['DelayCode_Description',
                  'DateAndTimeReturned',
                  'Postcode_full',
                  'Easting_m',
                  'Northing_m',
                  'Latitude',
                  'Longitude']
df_dropped = df_dropped.drop(labels = variables_drop2, axis = 1)

#Suppresion pour cause de colinéarité (valeurs identique par le calcul)
variables_drop3 =['DateOfCall',
                  'TimeOfCall',
                  'TurnoutTimeSeconds',
                  'TravelTimeSeconds',
                  'CalYear',
                  'HourOfCall']
df_dropped = df_dropped.drop(labels = variables_drop3, axis = 1)

df = df_dropped.copy()

## 02 - ajout des data météo

weather = pd.read_csv(racine_projet()+'/data/external/weather.csv', low_memory = False)

dataframe_info(weather)

df['DateAndTimeMobilised'] = pd.to_datetime(df['DateAndTimeMobilised'])
weather['date'] = pd.to_datetime(weather['date'])

# Troncature des colonnes datetime à l'heure
df['datetime_hour'] = df['DateAndTimeMobilised'].dt.floor('H')
weather['datetime_hour'] = weather['date'].dt.floor('H')

# S'assurer que les colonnes datetime_hour sont en UTC
df['datetime_hour'] = df['datetime_hour'].dt.tz_localize(None)  # Localize to None if not already UTC
weather['datetime_hour'] = weather['datetime_hour'].dt.tz_localize(None)

# Fusion des deux DataFrames en utilisant un left join
merged_df = pd.merge(df, weather, on='datetime_hour', how='left', suffixes=('_df1', '_df2'))

# Suppression des colonnes datetime_hour après la fusion
merged_df.drop(columns=['datetime_hour'], inplace=True)

dataframe_info(merged_df)

df = merged_df

## 02 bis - enregistrement dataviz (optionnel)

# Export du nouveau tableau sous format csv
df.to_csv(racine_projet()+'/data/processed/dataviz.csv', index=False)

## 03 - feature Engineering (après le dataviz)

df['DateAndTimeMobilised'] = pd.to_datetime(df['DateAndTimeMobilised'])

# Créer les colonnes 'year', 'month', 'weekday' et 'hour'
df['year'] = df['DateAndTimeMobilised'].dt.year
df['month'] = df['DateAndTimeMobilised'].dt.month
df['weekday'] = df['DateAndTimeMobilised'].dt.weekday  # Monday=0, Sunday=6
df['hour'] = df['DateAndTimeMobilised'].dt.hour

#Suppresion par inadaptabilité suite à DataViz
variables_dropfeats =['IncidentStationGround',
                      'FirstPumpArriving_DeployedFromStation',
                      'SecondPumpArriving_DeployedFromStation',
                      'DateAndTimeCalled',
                      'DateAndTimeMobilised',
                      'DateAndTimeMobile',
                      'DateAndTimeArrived',
                      'DateAndTimeLeft',
                      'AttendanceTimeSeconds',
                      'FirstPumpArriving_AttendanceTime',
                      'SecondPumpArriving_AttendanceTime',
                      'Postcode_district',
                      'IncGeo_WardName',
                      'IncGeo_WardNameNew',
                      'IncGeo_BoroughName',
                      'UPRN',
                      'USRN',
                      'IncidentGroup',
                      'StopCodeDescription',
                      'SpecialServiceType',
                      'PropertyType',
                      'PumpCount',
                      'PumpMinutesRounded',
                      'Notional Cost (£)',
                      'date'
                     ]
df_feats = df.drop(labels = variables_dropfeats, axis = 1)

dataframe_info(df_feats)

## 04 - traitement des NaN

df_cleaned = df_feats.copy()

# Suppression des lignes contenant des NaNs dans les features restantes : 'NumCalls', 'DeployedFromLocation','DeployedFromStation_Name'
#df_cleaned = df_cleaned.dropna(subset=['NumCalls'])
#df_cleaned = df_cleaned.dropna(subset=['DeployedFromLocation'])
#df_cleaned = df_cleaned.dropna(subset=['DeployedFromStation_Name'])
df_cleaned = df_cleaned.dropna()

dataframe_info(df_cleaned)

## 05 - cyclisation des variables cycliques (month/week/hours)

# Normalisation des variables numériques cycliques
df_norm = df_cleaned.copy()

# Utilisez .loc pour faire des assignations
df_norm['hour_sin'] = np.sin(2 * np.pi * df_norm['hour'] / 24)
df_norm['hour_cos'] = np.cos(2 * np.pi * df_norm['hour'] / 24)

df_norm['weekday_sin'] = np.sin(2 * np.pi * df_norm['weekday'] / 7)
df_norm['weekday_cos'] = np.cos(2 * np.pi * df_norm['weekday'] / 7)

df_norm.loc[:, 'month_sin'] = np.sin(2 * np.pi * df_norm['month'] / 12)
df_norm.loc[:, 'month_cos'] = np.cos(2 * np.pi * df_norm['month'] / 12)

df_norm = df_norm.drop(columns = ['month','weekday','hour'])

dataframe_info(df_norm)
dataframe_info(df_norm).to_csv(racine_projet()+'/data/processed/ML_data_info.csv', index=False)

## 06 - export

# Export du nouveau tableau sous format csv
df_norm.to_csv(racine_projet()+'/data/processed/ML_data.csv', index=False)
